% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_user_by_principal_id}
\alias{quicksight_delete_user_by_principal_id}
\title{Deletes a user after locating the user by its principal ID}
\usage{
quicksight_delete_user_by_principal_id(PrincipalId, AwsAccountId,
  Namespace)
}
\arguments{
\item{PrincipalId}{[required] The principal ID of the user.}

\item{AwsAccountId}{[required] The ID for the AWS account that the user is in. Currently, you use the
ID for the AWS account that contains your Amazon QuickSight account.}

\item{Namespace}{[required] The namespace. Currently, you should set this to \code{default}.}
}
\description{
Deletes a user after locating the user by its principal ID.
}
\section{Request syntax}{
\preformatted{quicksight$delete_user_by_principal_id(
  PrincipalId = "string",
  AwsAccountId = "string",
  Namespace = "string"
)
}
}

\keyword{internal}
